autotube/autotube/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
pydantic==2.9.2
duckduckgo-search==6.3.7
edge-tts==6.1.12
moviepy==1.0.3
Pillow==10.4.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
groq==0.11.0
